+++ src/main/scala/draco/drake/Drake.scala
@@ -8,7 +8,7 @@ object Drake extends App with DracoType {
   override lazy val typeDefinition: TypeDefinition = TypeLoader.loadType(TypeName ("Drake", _namePackage = Seq ("draco", "drake")))
   lazy val dracoType: Type[Drake] = Type[Drake] (typeDefinition)
 
-  lazy val elementTypeNames: Seq[String] = Seq ("Presence", "Present", "Absent")
+  lazy val elementTypeNames: Seq[String] = Seq ("Presence", "Present", "Absent", "Text")
 
   lazy val domainType: Domain[Drake] = Domain[Drake] (typeDefinition)
 }